Record symbols path in make
The Androidmk entries of a Soong module are processed in the mk file for
module types with a dedicated build rules mk file. `raw_binary` module
type does not have such build rules mk file; thus, the
INSTALLED_SYMBOLS.* variable is defined in the AndroidMkEntries method
of the `raw_binary` module type to record the symbols path, to ensure
that it does not conflict with that of the make prebuilt modules.

Test: lunch <conflicting product>-next-userdebug && m nothing
Bug: 427867754
Change-Id: I07196114a13eda4c1c0fcd1cad7967ca1b3b58ff