Add aconfig info to module_bp_java_deps.json
This CL adds a new `aconfig` field to java_aconfig_library modules in
module_bp_java_deps.json, which contains the srcs as well as the aconfig
configuration of the library so that aconfig files can be displayed and
recompiled in the IDE directly (e.g. using go/android-sonata).

Bug: 463647446
Test: go test java/
Test: go test android/
Test: go test aconfig/codegen/
Flag: EXEMPT HOST_MODE
Change-Id: Icccd6a84adc88ceea78428290412b0f5b9868891