Add tot_cronet_NetHttpCoverageTests to the exception list
`tot_cronet_NetHttpCoverageTests` is just a different naming for
`NetHttpCoverageTests`. However, we have to do things this way as we
have to synchronize across two different repos (Chromium and Android),
and doing everything in one go is painful. So just add
`tot_cronet_NetHttpCoverageTests` for the time-being and I'll come back
to remove `NetHttpCoverageTests` once the migration has finished.

Change-Id: I18c24984b6424a3b8fa27f9eef0bb04afd6c75e1