Add exportable output files for java_sdk_library
Previously, `java_sdk_library` only made the default (usually
`Everything`) files available as output files. That caused problems
for `frameworks-base-api-exportable-current.txt` which always needed to
be build from the `Exportable` files.

This change adds `exportableScopePaths` alongside `scopePaths` and
populates it with information extracted from the `Exportable` files
that the child `droidstubs` creates.

Future changes could use this to remove `exportableStubsDexJarPath` and
simplify the handling of the `ReleaseHiddenApiExportableStubs()`
configuration.

Bug: 425925746
Test: m checkapi
Change-Id: I1405e35c3530cc2e7369b548c2460d8f3b9859ee