Use zipsync in kati's extraFileZip install
The current unzip-based code fails with this error:
`is: -c: line 2: syntax error: unexpected end of file from `(' command on line 1`

Switch to zipsync to avoid compilicated shell logic.

We weren't seeing this issue before because this code only applies to
soong-only builds, and soong-only partitions use the CopySpecsToDirs()
function, instead of this code, which has already been updated to use
zipsync.

However, if you want to use `m mymodule`, this code will be run, causing
the error.

Bug: 452027180
Test: Built a prebuilt_kernel_modules module.
Change-Id: I8886e659be7c41e2af7e163627028bcba2f965a5