Remove unused Runtime and Unavailable StubsType
The `Runtime` value was never used, not specified in `Android.bp` files
or anywhere in the code. The `Unavailable` was only used to indicate an
error. This change removes the former and replaces the latter by
returning an `error` from `StringToStubsType`. The code was slightly
reorganized to clean that up.

Bug: 425925746
Test: m checkapi
Change-Id: Ic6eea7dc392e079e3caca6d7ce7942bdc5cd42af