Sandbox some python static rules
This sandboxes all the static rules except precompile, because
precompile depends on a cc_prebuilt_binary rule, which has the
problem described in b/444232435. We need to fix that first.

There are also 2 remaining unsandboxed Rulebuilder commands, and for
those we need support for having Rulebuilder commands that depend
on toybox tools, which I'll do in a followup.

Bug: 448484009
Test: SOONG_ACTION_SANDBOXING=nsjail m manifest_fixer
Change-Id: I34e82fea3c4b31da9f17c35beefe2a068086ad8e