Add generic filesystem prebuilt module
It was originally android_system_image_prebuilt but we've decided to
test system_ext prebuilt modules too. For now the draft implmenetation
is exactly the same, so this renames android_system_image_prebuilt to
android_filesystem_prebuilt.

Soong neverallows are still enforced so this never changes behavior for
normal targets. Note that the implementation is not final and this
implementation is to check any further blockers.

Bug: 461318449
Test: define system_ext prebuilt module and build
Change-Id: Id26d202b9e6195d5242e63595dd1ff65d47b88b9